+++ client/src/pages/StudentMyPage.js
@@ -1,4 +1,5 @@
 import React, { useState } from 'react';
+import { useNavigate, useParams } from 'react-router-dom';
 import axios from 'axios';
 
 import '../styles/studentMypage.scss';
@@ -7,16 +8,12 @@ import Template from '../components/Template';
 import { StudentIdCard } from '../components/StudentIdCard';
 import { PageHeader } from '../components/Headers';
 import { StudentPayStub } from './StudentPayStub';
-import { useParams } from 'react-router-dom';
 
 export default function StudentMyPage() {
-  const { id } = useParams();
-
-  const [isChangePassword, setIschangePassword] = useState(false);
-  const [newPassword, setNewPassword] = useState('');
+  const navigate = useNavigate();
 
   const handleChangePassword = () => {
-    setIschangePassword(true);
+    navigate('/:id/changePw');
   };
 
   return (
@@ -24,31 +21,10 @@ export default function StudentMyPage() {
       childrenTop={<PageHeader>{'마이페이지'}</PageHeader>}
       childrenBottom={
         <>
-          {isChangePassword ? (
-            <div className="changePassword-wrap">
-              <input
-                className="new-password"
-                type="password"
-                value={newPassword}
-                onChange={(e) => setNewPassword(e.target.value)}
-                placeholder="새 비밀번호"
-              />
-              <button className="change-btn">변경</button>
-              <button
-                className="cancelChange-btn"
-                onClick={() => setIschangePassword(false)}
-              >
-                취소
-              </button>
-            </div>
-          ) : (
-            <button
-              className="changePassword-btn"
-              onClick={handleChangePassword}
-            >
-              비밀번호 변경
-            </button>
-          )}
+          <button className="changePassword-btn" onClick={handleChangePassword}>
+            비밀번호 변경
+          </button>
+
           <StudentIdCard />
           <StudentPayStub />
         </>
