+++ client/src/styles/studentMypage.scss
@@ -43,7 +43,6 @@ $blue: #61759f;
     border-radius: 16px;
     padding: 20px;
     background-color: #fffcf2;
-    width: 280px;
     margin: 30px auto;
 
     .idCard-title {
@@ -94,92 +93,67 @@ $blue: #61759f;
   }
 
   // 월급 명세서
-  .payStub-wrap {
-    .payStub-title {
-      font-weight: 800;
+
+  .payStub-receipt {
+    border: 1px solid #b9c0be;
+    border-radius: 6px;
+    padding: 20px;
+    height: 300px;
+    overflow-x: hidden;
+    overflow-y: scroll;
+
+    .payStub-receipt-title {
+      font-size: 16px;
+      font-weight: 700;
+      margin-bottom: 20px;
     }
 
-    .payStub-month-navi {
+    .payStub-item {
+      margin-bottom: 20px;
+      display: flex;
+      flex-direction: column;
+
+      .payStub-img {
+        width: 16px;
+        margin-bottom: 6px;
+      }
+
+      .payStub-item-title,
+      .payStub-item-value {
+        font-size: 12px;
+        border-left: 2px solid #e2e8f0;
+        padding-left: 20px;
+        margin-left: 7px;
+      }
+    }
+
+    .payStub-total {
+      margin-top: 20px;
+      padding-top: 10px;
+      border-top: 1px solid;
       display: flex;
       align-items: center;
-      justify-content: space-between;
-      margin: 20px 0;
+      justify-content: flex-end;
 
-      .month-btn {
-        width: 30%;
-        font-size: 20px;
-        font-weight: 800;
-        border: 1px solid transparent;
-        background-color: transparent;
-      }
-
-      .month-btn:active {
-        box-shadow: 0px 0px 5px #444;
-      }
-    }
-
-    .payStub-receipt {
-      border: 1px solid #b9c0be;
-      border-radius: 6px;
-      padding: 20px;
-      height: 300px;
-      overflow-x: hidden;
-      overflow-y: scroll;
-
-      .payStub-receipt-title {
+      .total,
+      .total-pay {
         font-size: 14px;
-        font-weight: 700;
-        margin-bottom: 20px;
-      }
-
-      .payStub-item {
-        margin-bottom: 20px;
-        display: flex;
-        flex-direction: column;
-
-        .payStub-img {
-          width: 16px;
-          margin-bottom: 6px;
-        }
-
-        .payStub-item-title,
-        .payStub-item-value {
-          font-size: 12px;
-          border-left: 2px solid #e2e8f0;
-          padding-left: 20px;
-          margin-left: 7px;
-        }
-      }
-
-      .payStub-total {
-        margin-top: 20px;
-        padding-top: 10px;
-        border-top: 1px solid;
-        display: flex;
-        align-items: center;
-        justify-content: flex-end;
-
-        .total,
-        .total-pay,
-        .unit {
-          font-size: 14px;
-          margin-left: 10px;
-        }
+        margin-left: 10px;
       }
     }
+  }
 
-    .payStub-receipt::-webkit-scrollbar {
-      width: 10px;
-    }
+  .payStub-receipt::-webkit-scrollbar {
+    width: 10px;
+  }
 
-    .payStub-receipt::-webkit-scrollbar-thumb {
-      background: #d9d9d9; /* 스크롤바 색상 */
-      border-radius: 12px; /* 스크롤바 둥근 테두리 */
-    }
+  .payStub-receipt::-webkit-scrollbar-thumb {
+    background: #d9d9d9; /* 스크롤바 색상 */
+    border-radius: 12px; /* 스크롤바 둥근 테두리 */
+  }
 
-    .payStub-receipt::-webkit-scrollbar-track {
-      background: transparent; /*스크롤바 뒷 배경 색상*/
-      border-radius: 12px;
-    }
+  .payStub-receipt::-webkit-scrollbar-track {
+    background: transparent; /*스크롤바 뒷 배경 색상*/
+    border-radius: 12px;
   }
 }
